Robotic Hand control using Exo-Skeleton
This is a multi-disciplinary project. It includes mechanical skills,
along with electronic concepts as well as coding techniques. This
project is a small articulated arm controlled using potentiometer. Two
articulated arms were made. One of the arms had potentiometer at the
joints which was controlled by human and other had servo motors at the
joints. It is the property of a potentiometer to give variable
resistance when it is rotated. This change in resistance is converted
into angular motion in the controller which sets specific angle for the
servo motor to move simultaneously. The main use of this project is that
large machines can be controlled by the controller just by moving small
replica of the larger machines kept in the control room. This also helps
in accessing the places where human cannot enter.